+++ modules/engine/lib/engine.js
@@ -136,9 +136,11 @@ var Engine = module.exports = function(opts) {
         'mustache' : require('./serializers/mustache.js'),
         'ejs' : require('./serializers/ejs.js')
     }
-    this.cache = _util.getCache(this.config, opts.cache, function (err) {
+    this.cache = _util.getCache(this.config, opts.cache, self, function (err) {
         self.emitError('Unable to attach to Cache: ' + JSON.stringify(err, false, null));
     });
+
+
 }
 
 util.inherits(Engine, LogEmitter);