+++ CHANGES.md
@@ -1,3 +1,16 @@
+## May 3, 2012
+
+* Add support for multipart requests.
+
+## Apr 27, 2012
+
+* End pending connections on close after responses are written.
+* Support cache events (hit, miss, new, error, info, heartbeat)
+
+## Apr 26, 2012
+
+* Switch to new cluster2
+
 ## Apr 25, 2012
 
 * Added new syntaxes "with part" and opaque insert param.